Revert "Split executor test case file into 4 different files for cleanliness"
This reverts commit d73b1f65fae59bb79adaeda2770ee03671a91b43.
