+++ test/executor.js
@@ -322,4 +322,190 @@ contract('Executor', function(accounts) {
 
     });
 
+    describe("when service nodes process subscriptions", () => {
+
+        let oneMonthLater = activationTime + subscriptionInterval;
+
+        before(async function() {
+
+            // Set the time forward by one month
+            await executorContract.setTime(oneMonthLater);
+            await paymentRegistryContract.setTime(oneMonthLater);
+
+            // Transfer some enough tokens for cost and fee for the service node
+            await mockTokenContract.transfer(subscriber, subscriptionCost, + subscriptionFee, {from: contractOwner});
+
+            // Top up the user's account with funds for the next payment
+            await mockTokenContract.transfer(subscriber, subscriptionCost + fee, {from: contractOwner});
+
+        });
+
+        it("should not be able to process an inactive subscription", async function() {
+
+            // This will have a different hash since the timestamp is different
+            let inactiveSubscription = await subscriptionContract.createSubscription(
+                planHash, "{}", {from: subscriber}
+            );
+
+            // The hash we can use to identify the subscription
+            let inactiveSubscriptionHash = newSubscription.logs[0].args.identifier;
+
+            await assertRevert(executorContract.collectPayment(subscriptionContract.address, inactiveSubscriptionHash, {from: serviceNode}));
+
+        });
+
+        it("should not be able to process a subscription before the due date", async function() {
+
+            // Set the time right before the due date
+            await executorContract.turnBackTime(5);
+
+            await assertRevert(executorContract.collectPayment(subscriptionContract.address, subscriptionHash, {from: serviceNode}));
+
+            await executorContract.turnBackTime(-5);
+
+        });
+
+        it("should not be able to process a subscription if the service node does not have any staked tokens", async function() {
+
+            await assertRevert(executorContract.collectPayment(subscriptionContract.address, subscriptionHash, {from: serviceNode}));
+
+        });
+
+        it("should not be able to process a subscription if the service node does not have enough staked tokens", async function() {
+
+            // Transfer $10 * 10 (multiplier) worth of 8x tokens
+            let amount = (subscriptionCost * 10);
+            await nativeTokenContract.transfer(serviceNode,  amount, {from: contractOwner});
+
+            // Approve the stake contract to take 8x tokens from you
+            await nativeTokenContract.approve(stakeContract.address, amount, {from: serviceNode});
+
+            // Top up your stake in the stake contract less 100
+            await stakeContract.topUpStake(amount - 1000, {from: serviceNode});
+
+            // Check the balance to ensure it topped up the correct amount
+            let balance = await stakeContract.getAvailableStake(serviceNode.address);
+            assert.equal(balance, amount - 1000);
+
+            // Should fail when collecting payment since there aren't enough tokens
+            await assertRevert(executorContract.collectPayment(subscriptionContract.address, subscriptionHash, {from: serviceNode}));
+
+        });
+
+        it("should be able to process a valid subscription", async function() {
+
+            // Check to ensure the user has enough funds
+            let preUserBalance = await mockTokenContract.balanceOf(subscriber);
+            assert.equal(preUserBalance.toNumber(), subscriptionCost + subscriptionFee);
+
+            // Top up stake by difference
+            await stakeContract.topUpStake(1000, {from: serviceNode});
+
+            // Check the balance to ensure it topped up the correct amount
+            let balance = await stakeContract.getAvailableStake(serviceNode.address);
+            assert.equal(balance, amount);
+
+            // Should be able to process now that the service node has enough staked tokens
+            await executorContract.collectPayment(subscriptionContract.address, subscriptionHash, {from: serviceNode});
+
+            // Check the payments registry has been updated
+            let paymentInformation = await paymentRegistryContract.payments.call(subscriptionHash);
+
+            assert.equal(paymentInformation[0], subscriptionContract.address);
+            assert.equal(paymentInformation[1], oneMonthLater + subscriptionInterval);
+            assert.equal(paymentInformation[2], subscriptionCost);
+            assert.equal(paymentInformation[3], oneMonthLater);
+            assert.equal(paymentInformation[4], serviceNode);
+            assert.equal(paymentInformation[5], 0); // We claimed it straight away
+            assert.equal(paymentInformation[6], 10);
+
+            // Check if the service node got their reward, gas costs reimbursed and tokens locked up
+            // @TODO: Value for gas needs to be reimbursed
+            let serviceNodeStakeBalance = await stakeContract.getAvailableStake(serviceNode);
+            let serviceNodeTokenBalance = await mockTokenContract.balanceOf(serviceNode);
+
+            assert.equal(serviceNodeStakeBalance.toNumber(), 0);
+            assert.equal(serviceNodeTokenBalance.toNumber(), subscriptionFee);
+
+            // Check if the balance of the user was subtracted
+            let postUserBalance = await mockTokenContract.balanceOf(subscriber);
+            assert.equal(postUserBalance.toNumber(), 0);
+
+            // Check if the businesses' account was credited
+            let postBusinessBalance = await mockTokenContract.balanceOf(business);
+            assert.equal(postBusinessBalance.toNumber(), (subscriptionCost - fee) * 2);
+
+        });
+
+        it("should not be able to process a subscription once it has already been processed", async function() {
+
+            // Transfer $10 * 10 (multiplier) worth of 8x tokens
+            let amount = (subscriptionCost * 10);
+            await nativeTokenContract.transfer(competingServiceNode,  amount, {from: contractOwner});
+
+            // Approve the stake contract to take 8x tokens from you
+            await nativeTokenContract.approve(stakeContract.address, amount, {from: competingServiceNode});
+
+            // Top up your stake in the stake contract less 100
+            await stakeContract.topUpStake(amount, {from: competingServiceNode});
+
+            // Check the balance to ensure it topped up the correct amount
+            let balance = await stakeContract.getAvailableStake(competingServiceNode.address);
+            assert.equal(balance, amount);
+
+            // Should fail when collecting payment since it has already been claimed/processed
+            await assertRevert(executorContract.collectPayment(subscriptionContract.address, subscriptionHash, {from: competingServiceNode}));
+
+        });
+
+    });
+
+    describe("when service nodes cancel a subscription", () => {
+
+        it("should not be able to if the due date has already passed", async function() {
+
+            // @TODO: Implementation
+
+        });
+
+        it("should be able to before the due date", async function() {
+
+            // @TODO: Implementation
+
+        });
+
+    });
+
+    describe("when businesses cancel a subscription", () => {
+
+        it("should not be able to cancel to an unauthorised subscription contract", async function() {
+
+            // @TODO: Protocol
+
+        });
+
+        it("should not be able to as an unauthorised user", async function() {
+
+            // @TODO: Implementation
+
+        });
+
+    });
+
+    describe("when users don't have enough funds in their wallet", () => {
+
+        it("should not allow the subscription to be processed", async function() {
+
+            // @TODO: Implementation
+
+        });
+
+        it("should cancel the subscription", async function() {
+
+            // @TODO: Implementation
+
+        });
+
+    });
+
 });
